requirements.txt

torch>=2.2.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.22.0
peft>=0.6.0
bitsandbytes>=0.41.0
trl>=0.9.0
sentencepiece>=0.1.99
streamlit>=1.28.0
nltk>=3.8.1
PyYAML>=6.0